requirements.txt
ioconsole==0.4.1
aiofiles==0.8.0  
aiohttp==3.8.1   
aioopenssl==0.6.0
aiosasl==0.5.0   
aiosignal==1.2.0 
aioxmpp==0.13.2  
altgraph==0.17.2
antiorm==1.2.1
asgiref==3.5.2
astc-decomp==1.0.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.1
beautifulsoup4==4.10.0
BenBotAsync==3.0.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.5
crayons==0.4.0
cryptography==3.2.1
Cython==0.29.32
DateTime==4.5
db==0.1.1
discord-py-interactions==4.2.1
discord-py-slash-command==4.2.1
discord-ui==5.1.6
discord.py==1.7.3
Django==4.0.5
django-ipware==4.0.2
dnspython==2.2.1
ffmpeg-python==0.2.0
Flask==2.1.2
FortniteAPIAsync==0.1.6
fortnitepy==3.6.5
frozenlist==1.3.0
future==0.18.2
gevent==21.12.0
glob2==0.7
greenlet==1.1.2
httptools==0.4.0
idna==3.3
image==1.5.33
imageio==2.19.3
imageio-ffmpeg==0.4.7
install==1.3.5
instaloader==4.9.3
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.71
legendary-gl==0.20.27
loguru==0.6.0
lxml==4.8.0
lz4==4.0.2
MarkupSafe==2.1.1
multidict==6.0.2
nextcord==2.1.0
numpy==1.22.4
oauthlib==3.2.0
palettable==3.3.0
pandas==1.4.3
path==16.4.0
pefile==2022.5.30
Pillow==9.2.0
proxy-tools==0.1.0
psutil==5.9.0
py-cord @ git+https://github.com/Pycord-Development/pycord@a2b1d8f8316c133ff58fa7928d1a7c36948a8f79
py-wallpaper==0.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycord==0.1.1
pycparser==2.21
pycryptodome==3.15.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyjsparser==2.7.1
PyNaCl==1.5.0
pyOpenSSL==20.0.1
pypresence==4.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytube==12.1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyUsmap==1.1.3
pywin32-ctypes==0.2.0
quicktex==0.1.3
replit==3.2.4
requests==2.28.1
requests-oauthlib==1.3.1
sanic==22.3.2
sanic-ipware==0.1.0
sanic-routing==22.3.0
SClib==1.0.0
six==1.16.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
stem==1.8.0
toml==0.10.2
tweepy==4.10.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
UE4Parse==0.0.1
urllib3==1.26.9
wallpaper==0.2.2
websocket==0.2.1
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.1.2
win32-setctime==1.1.0
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0